move /lib/autovalidate -> /lib/form-view/autovalidate
