+++ lib/form-view/form-view.js
@@ -3,7 +3,7 @@ import spin from 'component-spin'
 import t from 't-component'
 import * as validate from '../validate/validate.js'
 import autosubmit from './autosubmit/autosubmit'
-import autovalidate from '../autovalidate/autovalidate.js'
+import autovalidate from './autovalidate/autovalidate'
 import isMobile from '../is-mobile/is-mobile.js'
 import View from '../view/view.js'
